requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.1.0
langchain-openai>=0.0.2
pydantic>=2.0.0
click>=8.0.0
colorama>=0.4.6
rich>=13.0.0
python-dotenv>=0.19.0
requests>=2.31.0
prompt_toolkit>=3.0.0
beautifulsoup4>=4.12.0
psutil>=5.9.0